Create new default package Skylark
