+++ src/skylark/skylark/Deadline.java
@@ -1,3 +1,5 @@
+package skylark;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
@@ -6,7 +8,6 @@ public class Deadline extends Task {
     private final char SYMBOL = 'D';
 
     private final String inputFormat = "yyyy-MM-dd HHmm"; // 2019-10-15 1800
-    private final String outputFormat = "MMM dd yyyy"; // Oct 15 2019
     private final LocalDateTime endDate;
 
     public Deadline(String description, String endDate) throws SkylarkException {
@@ -20,6 +21,8 @@ public class Deadline extends Task {
 
     @Override
     public String toString() {
+        // Oct 15 2019
+        String outputFormat = "MMM dd yyyy";
         return String.format("[%c] %s (by: %s)",
                 SYMBOL, super.toString(), this.endDate.format(DateTimeFormatter.ofPattern(outputFormat)));
     }
@@ -27,7 +30,7 @@ public class Deadline extends Task {
     @Override
     public String toStringFile() {
         return String.format("%c | %d | %s | %s", SYMBOL,
-                super.getStatusIcon() == "X" ? 1 : 0, super.getDescription(),
+                super.getStatusIcon().equals("X") ? 1 : 0, super.getDescription(),
                 this.endDate.format(DateTimeFormatter.ofPattern(inputFormat)));
     }
 }
