+++ src/skylark/skylark/DeadlineException.java
@@ -1,0 +1,16 @@
+package skylark;
+
+public class DeadlineException extends SkylarkException{
+
+    private final String description;
+
+    public DeadlineException(String description) {
+        super(description);
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return this.description;
+    }
+}
