+++ src/skylark/skylark/Event.java
@@ -1,3 +1,5 @@
+package skylark;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
@@ -6,7 +8,6 @@ public class Event extends Task{
     private final char SYMBOL = 'E';
 
     private final String inputFormat = "yyyy-MM-dd HHmm"; // 2019-10-15 1800
-    private final String outputFormat = "MMM dd yyyy"; // Oct 15 2019
     private final LocalDateTime timing;
 
     public Event(String description, String timing) throws SkylarkException {
@@ -20,6 +21,8 @@ public class Event extends Task{
 
     @Override
     public String toString() {
+        // Oct 15 2019
+        String outputFormat = "MMM dd yyyy";
         return String.format("[%c] %s (at: %s)", SYMBOL,
                 super.toString(), this.timing.format(DateTimeFormatter.ofPattern(outputFormat)));
     }
@@ -27,7 +30,7 @@ public class Event extends Task{
     @Override
     public String toStringFile() {
         return String.format("%c | %d | %s | %s", SYMBOL,
-                super.getStatusIcon() == "X" ? 1 : 0,
+                super.getStatusIcon().equals("X") ? 1 : 0,
                 super.getDescription(), this.timing.format(DateTimeFormatter.ofPattern(inputFormat)));
     }
 }
