+++ src/skylark/skylark/EventException.java
@@ -1,6 +1,8 @@
+package skylark;
+
 public class EventException extends SkylarkException{
 
-    private String description;
+    private final String description;
 
     public EventException(String description) {
         super(description);
